Task 1 (building crawler) finished.
The crawler has been built and works. Several classes have been built to
the project a basic structure. The Crawler class does crawl through the
net and saves the found pages in the Pages class. It keeps track of how
	many pages and links have been encountered.
The Page class represents the relevant information of the fetched pages
in a structured way. The Parser Class provides a static function to
convert the data from a fetched site to Page object.