Task 2 completed. Indexer Class added
The Class Indexer has been added which handles the forward and
reverse index. The Class uses the Crawler class to get a list of
crawled Pages. This structure has been chosen to represent the
structure Fig 3.2.
The Indexer contains the forward and reverse index as maps, since not
much more functionality is needed so far.
The unit tests have been implemented and have been passed.