Added PageRank Class and modified some Classes
Changed attribute of Page Class from linkList to linkSet, since it is
not necessary to save multiple outgoing links to the same Page. This
change makes implementation of the PageRank algorithm a bit simpler,
since there is no need to check for duplicates there.
PageRank algorithm implemented as a class. No dampening implemented so
far. Testing has shown that it oscilates/loops, when testet on cheese6
net.