Pagerank class implemented dampening
The pagerank calulation has been modified to include a ranksource and
dampening.
The unit test has been modified accordingly and is being passed.
A new ranking method has been implemented combining the cosine
similarity and pagerank. Since the absolute pagerank value is highly
dependant on the amount of pages crawled both values have been
normalized to be a factor of the average of the search results. The
influence of the page rank should independant of the total crawled pages
with this normalization.