Main program added in AnguillaSearch class
A main program has been added, where it is possible to search for terms.
It is possible to specify some programs aspects via arguments. Currently
these are implemented.
-r <number> select ranking method for ranking search results.
-color enables ANSI color highlighting for search results.
The last argument can be used to specify a path to a json file or seed
URLs.
The parser still doesn't work properly and does't parse the arguments
-color myjsonfile.json correctly.