Added funcionality to weight tokens
For the cosine similarity ranking now supports weight. Has been added as
ranking method 3. If a token is multiple times an a searchterm this will
change the weight accordingly.
Since all ranking methods have been using sets of search tokens the
implementation of the weighting is right now in the searchQuery
function, which only selects the ranking method. Might change this
later.