Bugfix and added normalization of tfidf-vectors
Removed some code which has been used for testing purposes and caused an
infinite loop.
When creating the tfidf-vectors in the forward index all vector are
normalized to a length of 1. The search vector used when ranking using
the cosine similarity method is normalized, too.
The divisor in the cosine similarity becomes 1 * 1 and therefor the
euclidian norm calculation can be skipped.